+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 2.11.0
 ### Features
 * improve performance
+* add cache for formatting and processing lines
+* big improvements to less plugin re-rendering (e.g. when scrolling text with keys)
 ### Bugfix
 * fix empty lines in less
 * fix split_equal with keep works and formatting at the end.